Update README.md
Hello, I'm learning how to code. This is the first time ever I've tried to do it. I have a financial background but I feel that coding is my true calling. I want to learn this new skill as fast as possible and hopefully we could work together sometime in the future.